# packages/shiori-api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "shiori",
    "description": "TODO DESCRIPTION OF SHIORI",
    "license": {
      "name": ""
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/media/{id}": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Fetch media item.",
        "operationId": "get_media",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the media item",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched media cover",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Media"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "metadata": {
                          "oneOf": [
                            {
                              "type": "null"
                            },
                            {
                              "$ref": "#/components/schemas/MediaMetadata"
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Media not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "media"
        ],
        "summary": "Delete a media item.",
        "operationId": "delete_media",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the media item",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully delete media"
          },
          "404": {
            "description": "Media not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "tags": [
          "media"
        ],
        "summary": "Update media information.",
        "operationId": "patch_media",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the media item",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cover_url": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "description": "URL of the cover image associated with the media.",
                    "examples": [
                      "https://example.com/cover.jpg"
                    ]
                  },
                  "metadata": {
                    "oneOf": [
                      {
                        "type": "null"
                      },
                      {
                        "$ref": "#/components/schemas/PatchMetadata",
                        "description": "Optional metadata to update for the media item."
                      }
                    ]
                  },
                  "name": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "description": "Name of the media item.",
                    "examples": [
                      "86—EIGHTY-SIX, Vol. 1"
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully updated media information",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Media"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "metadata": {
                          "oneOf": [
                            {
                              "type": "null"
                            },
                            {
                              "$ref": "#/components/schemas/MediaMetadata"
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Media not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/media/{id}/cover": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Fetch media cover.",
        "operationId": "get_media_cover",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the media item",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched media cover"
          },
          "404": {
            "description": "Media or cover not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Media": {
        "type": "object",
        "required": [
          "id",
          "name",
          "size",
          "path",
          "extension",
          "created_at",
          "library_id"
        ],
        "properties": {
          "cover_path": {
            "type": [
              "string",
              "null"
            ],
            "description": "Endpoint where the cover is stored.",
            "examples": [
              "/api/v1/media/4/cover"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of when the media was created.",
            "examples": [
              "2026-03-23T12:00:00Z"
            ]
          },
          "extension": {
            "type": "string",
            "description": "File extension of the media.",
            "examples": [
              "epub"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "Unique identifier for the media item.",
            "examples": [
              86
            ]
          },
          "library_id": {
            "type": "integer",
            "format": "int32",
            "description": "Library this media belongs to.",
            "examples": [
              2
            ]
          },
          "name": {
            "type": "string",
            "description": "Name of the media file, excluding extension.",
            "examples": [
              "86_Volume_1"
            ]
          },
          "path": {
            "type": "string",
            "description": "File system path where the media is stored.",
            "examples": [
              "/data/books/light_novels/86_Volume_1.epub"
            ]
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "description": "Size of the media file in bytes.",
            "examples": [
              "102400"
            ]
          }
        }
      },
      "MediaMetadata": {
        "type": "object",
        "required": [
          "authors",
          "genres"
        ],
        "properties": {
          "authors": {
            "type": "array",
            "items": {
              "type": "string",
              "examples": [
                [
                  "Asato Asato"
                ]
              ]
            },
            "description": "List of authors associated with the media item."
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Description of the media item.",
            "examples": [
              "The San Magnolia Republic..."
            ]
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string",
              "examples": [
                [
                  "Light Novel",
                  "War"
                ]
              ]
            },
            "description": "List of genres associated with the media item."
          },
          "isbn": {
            "type": [
              "string",
              "null"
            ],
            "description": "International Standard Book Number (ISBN).\nTypically used for books.",
            "examples": [
              "1975303121"
            ]
          },
          "language": {
            "type": [
              "string",
              "null"
            ],
            "description": "Language of the media content.",
            "examples": [
              "English"
            ]
          },
          "published_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date",
            "description": "Date the media was published.",
            "examples": [
              "2019-03-26"
            ]
          },
          "publisher": {
            "type": [
              "string",
              "null"
            ],
            "description": "Name of the publisher or publishing organization.",
            "examples": [
              "Yen On"
            ]
          }
        }
      },
      "PatchMetadata": {
        "type": "object",
        "properties": {
          "authors": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string",
              "examples": [
                [
                  "Asato Asato"
                ]
              ]
            },
            "description": "List of authors associated with the media item."
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Description of the media item.",
            "examples": [
              "The San Magnolia Republic..."
            ]
          },
          "genres": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string",
              "examples": [
                [
                  "Light Novel",
                  "War"
                ]
              ]
            },
            "description": "List of genres associated with the media item."
          },
          "isbn": {
            "type": [
              "string",
              "null"
            ],
            "description": "International Standard Book Number (ISBN).\nTypically used for books.",
            "examples": [
              "1975303121"
            ]
          },
          "language": {
            "type": [
              "string",
              "null"
            ],
            "description": "Language of the media content.",
            "examples": [
              "English"
            ]
          },
          "published_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date",
            "description": "Date the media was published.",
            "examples": [
              "2019-03-26"
            ]
          },
          "publisher": {
            "type": [
              "string",
              "null"
            ],
            "description": "Name of the publisher or publishing organization.",
            "examples": [
              "Yen On"
            ]
          }
        }
      }
    }
  }
}